+++ c/array-fpi/ssina.c
@@ -52,7 +52,7 @@ int main()
 
 	for(i=0; i<N; i++)
 	{
-		__VERIFIER_assert(a[i] == (N+1) * (N+1));
+		__VERIFIER_assert(a[i] == (long long)(N+1) * (N+1));
 	}
 	return 1;
 }
